refactor: merge ProteinIdentificationArrowExport/Import into ProteinIdentificationArrowIO
Combine the separate Export and Import classes into a single
ProteinIdentificationArrowIO class to simplify the API surface.
Deduplicates anonymous namespace helpers and updates all cmake
files, tests, and Python bindings accordingly.
