+++ src/openms/source/FORMAT/sources.cmake
@@ -112,8 +112,7 @@ if (WITH_PARQUET)
   list(APPEND sources_list ParquetFilter.cpp)
   list(APPEND sources_list QPXFile.cpp)
   list(APPEND sources_list XICParquetFile.cpp)
-  list(APPEND sources_list ProteinIdentificationArrowExport.cpp)
-  list(APPEND sources_list ProteinIdentificationArrowImport.cpp)
+  list(APPEND sources_list ProteinIdentificationArrowIO.cpp)
 endif()
 
 ### add path to the filenames
