+++ src/tests/class_tests/openms/executables.cmake
@@ -290,8 +290,7 @@ if(WITH_PARQUET)
   list(APPEND format_executables_list Arrow_test MSExperimentArrowExport_test ConsensusMapArrowExport_test QPXFile_test
     MSChromatogramParquetConsumer_test
     XICParquetFile_test
-    ProteinIdentificationArrowExport_test
-    ProteinIdentificationArrowImport_test)
+    ProteinIdentificationArrowIO_test)
 endif()
 
 set(math_executables_list
